# =str Optimize performance of MapAsyncUnordered(1).