--- a/akka-stream-tests/src/test/scala/akka/stream/scaladsl/FlowMapAsyncUnordered1Spec.scala
+++ b/akka-stream-tests/src/test/scala/akka/stream/scaladsl/FlowMapAsyncUnordered1Spec.scala
@@ -1,20 +1,8 @@
 /*
- * Copyright (C) 2014-2023 Lightbend Inc. <https://www.lightbend.com>
+ * Copyright (C) 2014-2022 Lightbend Inc. <https://www.lightbend.com>
  */
 
 package akka.stream.scaladsl
-
-import java.util.concurrent.LinkedBlockingQueue
-import java.util.concurrent.atomic.AtomicInteger
-
-import scala.annotation.tailrec
-import scala.concurrent.Await
-import scala.concurrent.Future
-import scala.concurrent.Promise
-import scala.concurrent.duration._
-import scala.util.control.NoStackTrace
-
-import org.scalatest.concurrent.PatienceConfiguration.Timeout
 
 import akka.stream.ActorAttributes.supervisionStrategy
 import akka.stream.Supervision.resumingDecider
@@ -23,7 +11,12 @@ import akka.stream.testkit.scaladsl._
 import akka.testkit.TestLatch
 import akka.testkit.TestProbe
 
-class FlowMapAsyncUnorderedSpec extends StreamSpec {
+import scala.concurrent.Await
+import scala.concurrent.Future
+import scala.concurrent.duration._
+import scala.util.control.NoStackTrace
+
+class FlowMapAsyncUnordered1Spec extends StreamSpec {
 
   "A Flow with mapAsyncUnordered" must {
 
@@ -32,7 +25,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
       implicit val ec = system.dispatcher
       val latch = (1 to 4).map(_ -> TestLatch(1)).toMap
       Source(1 to 4)
-        .mapAsyncUnordered(4)(n =>
+        .mapAsyncUnordered(1)(n =>
           Future {
             Await.ready(latch(n), 5.seconds)
             n
@@ -41,47 +34,47 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
         .run()
       val sub = c.expectSubscription()
       sub.request(5)
-      latch(2).countDown()
-      c.expectNext(2)
-      latch(4).countDown()
-      c.expectNext(4)
-      latch(3).countDown()
-      c.expectNext(3)
       latch(1).countDown()
-      c.expectNext(1)
+      c.expectNext(6.seconds, 1)
+      latch(2).countDown()
+      c.expectNext(6.seconds, 2)
+      latch(3).countDown()
+      c.expectNext(6.seconds, 3)
+      latch(4).countDown()
+      c.expectNext(6.seconds, 4)
       c.expectComplete()
     }
 
-    "complete without requiring further demand (parallelism = 2)" in {
+    "complete without requiring further demand (parallelism = 1)" in {
       import system.dispatcher
       Source
         .single(1)
-        .mapAsyncUnordered(2)(v => Future { Thread.sleep(20); v })
+        .mapAsyncUnordered(1)(v => Future { Thread.sleep(20); v })
         .runWith(TestSink[Int]())
         .requestNext(1)
         .expectComplete()
     }
 
-    "complete without requiring further demand with already completed future (parallelism = 2)" in {
+    "complete without requiring further demand with already completed future (parallelism = 1)" in {
       Source
         .single(1)
-        .mapAsyncUnordered(2)(v => Future.successful(v))
+        .mapAsyncUnordered(1)(v => Future.successful(v))
         .runWith(TestSink[Int]())
         .requestNext(1)
         .expectComplete()
     }
 
-    "complete without requiring further demand (parallelism = 2, multi value, multi values source)" in {
+    "complete without requiring further demand (parallelism = 1) with multi values" in {
       import system.dispatcher
       val probe =
-        Source(1 :: 2 :: Nil).mapAsyncUnordered(2)(v => Future { Thread.sleep(20); v }).runWith(TestSink[Int]())
+        Source(1 :: 2 :: Nil).mapAsyncUnordered(1)(v => Future { Thread.sleep(20); v }).runWith(TestSink[Int]())
 
       probe.request(2).expectNextN(2)
       probe.expectComplete()
     }
 
-    "complete without requiring further demand with already completed future (parallelism = 2, with multi values)" in {
-      val probe = Source(1 :: 2 :: Nil).mapAsyncUnordered(2)(v => Future.successful(v)).runWith(TestSink[Int]())
+    "complete without requiring further demand with already completed future (parallelism = 1) with multi values" in {
+      val probe = Source(1 :: 2 :: Nil).mapAsyncUnordered(1)(v => Future.successful(v)).runWith(TestSink[Int]())
 
       probe.request(2).expectNextN(2)
       probe.expectComplete()
@@ -92,7 +85,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
       val c = TestSubscriber.manualProbe[Int]()
       implicit val ec = system.dispatcher
       Source(1 to 20)
-        .mapAsyncUnordered(4)(n =>
+        .mapAsyncUnordered(1)(n =>
           if (n % 3 == 0) {
             probe.ref ! n
             Future.successful(n)
@@ -108,10 +101,10 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
       probe.expectNoMessage(Duration.Zero)
       sub.request(1)
       var got = Set(c.expectNext())
-      probe.expectMsgAllOf(1, 2, 3, 4, 5)
+      probe.expectMsg(1)
       probe.expectNoMessage(500.millis)
-      sub.request(25)
-      probe.expectMsgAllOf(6 to 20: _*)
+      sub.request(19)
+      probe.expectMsgAllOf(2 to 20: _*)
       c.within(3.seconds) {
         for (_ <- 2 to 20) got += c.expectNext()
       }
@@ -121,15 +114,13 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
     }
 
     "signal future failure" in {
-      val latch = TestLatch(1)
       val c = TestSubscriber.manualProbe[Int]()
       implicit val ec = system.dispatcher
       Source(1 to 5)
-        .mapAsyncUnordered(4)(n =>
+        .mapAsyncUnordered(1)(n =>
           Future {
             if (n == 3) throw new RuntimeException("err1") with NoStackTrace
             else {
-              Await.ready(latch, 10.seconds)
               n
             }
           })
@@ -137,8 +128,9 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
         .run()
       val sub = c.expectSubscription()
       sub.request(10)
+      c.expectNext(1)
+      c.expectNext(2)
       c.expectError().getMessage should be("err1")
-      latch.countDown()
     }
 
     "signal future failure asap" in {
@@ -152,7 +144,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
             n
           }
         }
-        .mapAsyncUnordered(4) { n =>
+        .mapAsyncUnordered(1) { n =>
           if (n == 1) Future.failed(new RuntimeException("err1") with NoStackTrace)
           else Future.successful(n)
         }
@@ -163,31 +155,10 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
       latch.countDown()
     }
 
-    "signal error from mapAsyncUnordered" in {
-      val latch = TestLatch(1)
-      val c = TestSubscriber.manualProbe[Int]()
-      implicit val ec = system.dispatcher
-      Source(1 to 5)
-        .mapAsyncUnordered(4)(n =>
-          if (n == 3) throw new RuntimeException("err2") with NoStackTrace
-          else {
-            Future {
-              Await.ready(latch, 10.seconds)
-              n
-            }
-          })
-        .to(Sink.fromSubscriber(c))
-        .run()
-      val sub = c.expectSubscription()
-      sub.request(10)
-      c.expectError().getMessage should be("err2")
-      latch.countDown()
-    }
-
     "resume after future failure" in {
       implicit val ec = system.dispatcher
       Source(1 to 5)
-        .mapAsyncUnordered(4)(n =>
+        .mapAsyncUnordered(1)(n =>
           Future {
             if (n == 3) throw new RuntimeException("err3") with NoStackTrace
             else n
@@ -210,7 +181,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
 
       Await.result(
         Source(futures)
-          .mapAsyncUnordered(2)(identity)
+          .mapAsyncUnordered(1)(identity)
           .withAttributes(supervisionStrategy(resumingDecider))
           .runWith(Sink.head),
         3.seconds) should ===("happy!")
@@ -220,22 +191,21 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
       import system.dispatcher
       Await.result(
         Source(1 to 3)
-          .mapAsyncUnordered(2)(n =>
+          .mapAsyncUnordered(1)(n =>
             Future {
               if (n == 3) throw new RuntimeException("err3b") with NoStackTrace
               else n
             })
           .withAttributes(supervisionStrategy(resumingDecider))
           .grouped(10)
-          .map(_.toSet)
           .runWith(Sink.head),
-        1.second) should be(Set(1, 2))
+        1.second) should be(Seq(1, 2))
     }
 
     "resume when mapAsyncUnordered throws" in {
       implicit val ec = system.dispatcher
       Source(1 to 5)
-        .mapAsyncUnordered(4)(n =>
+        .mapAsyncUnordered(1)(n =>
           if (n == 3) throw new RuntimeException("err4") with NoStackTrace
           else Future(n))
         .withAttributes(supervisionStrategy(resumingDecider))
@@ -246,7 +216,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
     }
 
     "ignore element when future is completed with null" in {
-      val flow = Flow[Int].mapAsyncUnordered[String](2) {
+      val flow = Flow[Int].mapAsyncUnordered[String](1) {
         case 2 => Future.successful(null)
         case x => Future.successful(x.toString)
       }
@@ -256,7 +226,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
     }
 
     "continue emitting after a sequence of nulls" in {
-      val flow = Flow[Int].mapAsyncUnordered[String](3) { value =>
+      val flow = Flow[Int].mapAsyncUnordered[String](1) { value =>
         if (value == 0 || value >= 100) Future.successful(value.toString)
         else Future.successful(null)
       }
@@ -267,7 +237,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
     }
 
     "complete without emitting any element after a sequence of nulls only" in {
-      val flow = Flow[Int].mapAsyncUnordered[String](3) { _ =>
+      val flow = Flow[Int].mapAsyncUnordered[String](1) { _ =>
         Future.successful(null)
       }
 
@@ -278,16 +248,13 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
 
     "complete stage if future with null result is completed last" in {
       import system.dispatcher
-      val latch = TestLatch(2)
 
-      val flow = Flow[Int].mapAsyncUnordered[String](2) {
+      val flow = Flow[Int].mapAsyncUnordered[String](1) {
         case 2 =>
           Future {
-            Await.ready(latch, 10.seconds)
             null
           }
         case x =>
-          latch.countDown()
           Future.successful(x.toString)
       }
 
@@ -300,7 +267,7 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
       val pub = TestPublisher.manualProbe[Int]()
       val sub = TestSubscriber.manualProbe[Int]()
 
-      Source.fromPublisher(pub).mapAsyncUnordered(4)(Future.successful).runWith(Sink.fromSubscriber(sub))
+      Source.fromPublisher(pub).mapAsyncUnordered(1)(Future.successful).runWith(Sink.fromSubscriber(sub))
 
       val upstream = pub.expectSubscription()
       upstream.expectRequest()
@@ -309,52 +276,6 @@ class FlowMapAsyncUnorderedSpec extends StreamSpec {
 
       upstream.expectCancellation()
 
-    }
-
-    "not run more futures than configured" in {
-      val parallelism = 8
-
-      val counter = new AtomicInteger
-      val queue = new LinkedBlockingQueue[(Promise[Int], Long)]
-
-      val timer = new Thread {
-        val delay = 50000 // nanoseconds
-        var count = 0
-        @tailrec final override def run(): Unit = {
-          val cont = try {
-            val (promise, enqueued) = queue.take()
-            val wakeup = enqueued + delay
-            while (System.nanoTime() < wakeup) {}
-            counter.decrementAndGet()
-            promise.success(count)
-            count += 1
-            true
-          } catch {
-            case _: InterruptedException => false
-          }
-          if (cont) run()
-        }
-      }
-      timer.start
-
-      def deferred(): Future[Int] = {
-        if (counter.incrementAndGet() > parallelism) Future.failed(new Exception("parallelism exceeded"))
-        else {
-          val p = Promise[Int]()
-          queue.offer(p -> System.nanoTime())
-          p.future
-        }
-      }
-
-      try {
-        val N = 10000
-        Source(1 to N)
-          .mapAsyncUnordered(parallelism)(_ => deferred())
-          .runFold(0)((c, _) => c + 1)
-          .futureValue(Timeout(3.seconds)) should ===(N)
-      } finally {
-        timer.interrupt()
-      }
     }
 
   }