Run with 
> jmh:run -i 20 -wi 20 -f1 -t1 .*MapAsyncBenchmark.*
I adjusted it a little to just run with `1`.

>jmh:run -i 20 -wi 20 -f1 -t1 .MapAsyncOneBenchmark.

Before:
```
[info] Iteration  20: 2022-09-11 17:14:30,867 INFO  akka.actor.CoordinatedShutdown  - Running CoordinatedShutdown with reason [ActorSystemTerminateReason] MDC: {akkaAddress=akka://MapAsyncBenchmark, akkaUid=-6827262736314033361, sourceThread=akka.stream.MapAsyncBenchmark.mapAsyncUnordered-jmh-worker-1, akkaSource=CoordinatedShutdown(akka://MapAsyncBenchmark), sourceActorSystem=MapAsyncBenchmark, akkaTimestamp=09:14:30.864UTC}
[info] 306857.968 ops/s
[info] Result "akka.stream.MapAsyncBenchmark.mapAsyncUnordered":
[info]   319643.963 锟斤拷(99.9%) 25725.842 ops/s [Average]
[info]   (min, avg, max) = (282843.475, 319643.963, 363219.309), stdev = 29625.918
[info]   CI (99.9%): [293918.121, 345369.805] (assumes normal distribution)
[info] # Run complete. Total time: 00:13:31
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                            (parallelism)  (spawn)   Mode  Cnt         Score        Error  Units
[info] MapAsyncBenchmark.mapAsyncUnordered              1    false  thrpt   20  15452906.430 锟斤拷 244116.496  ops/s
[info] MapAsyncBenchmark.mapAsyncUnordered              1     true  thrpt   20    319643.963 锟斤拷  25725.842  ops/s
```

after 

```
[info] Iteration  20: 2022-09-11 17:31:28,159 INFO  akka.actor.CoordinatedShutdown  - Running CoordinatedShutdown with reason [ActorSystemTerminateReason] MDC: {akkaAddress=akka://MapAsyncBenchmark, akkaUid=-5508069858145503529, sourceThread=akka.stream.MapAsyncBenchmark.mapAsyncUnordered-jmh-worker-1, akkaSource=CoordinatedShutdown(akka://MapAsyncBenchmark), sourceActorSystem=MapAsyncBenchmark, akkaTimestamp=09:31:28.157UTC}
[info] 356229.926 ops/s
[info] Result "akka.stream.MapAsyncBenchmark.mapAsyncUnordered":
[info]   341701.873 锟斤拷(99.9%) 21521.068 ops/s [Average]
[info]   (min, avg, max) = (288813.936, 341701.873, 367134.275), stdev = 24783.694
[info]   CI (99.9%): [320180.805, 363222.940] (assumes normal distribution)
[info] # Run complete. Total time: 00:13:31
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                            (parallelism)  (spawn)   Mode  Cnt         Score        Error  Units
[info] MapAsyncBenchmark.mapAsyncUnordered              1    false  thrpt   20  19811428.492 锟斤拷 119502.131  ops/s
[info] MapAsyncBenchmark.mapAsyncUnordered              1     true  thrpt   20    393896.158 锟斤拷 19592.226  ops/s


```