@He-Pin 2022-09-14
I tried to parameterized the spec but there are tests only written for parallelism 1 or 4, and the result code not that clear too