@johanandren 2022-09-14
Taking a look at the MapAsyncUnordered impl, when size is 1 I'd expect that it does not actually go via buffer and be similar to your optimized version where it just pushes immediately both when directly completed and completed later. Curious.