@johanandren 2022-09-14
Maybe, but for FixedBuffer those mostly boil down to integer ops, so a 30% improvement seems suspicious to me.