@He-Pin 2022-09-14
I ran it at windows 11 with 16cpu 4gb jvm.