@johanandren 2022-09-14
On my MBP M1 Max (arm64), JDK 11 I cannot reproduce your results, I did a few different runs with pretty unstable results with, somewhat what I would expect for high throughput and roughly equivalent logic. Here is one result (fewer iterations but results were pretty stable even earlier than 5 warmups):

```
[info] Benchmark                             (parallelism)  (spawn)   Mode  Cnt         Score        Error  Units
[info] MapAsyncBenchmark.mapAsyncUnordered               1    false  thrpt    5  19616888,260 ± 105909,078  ops/s
[info] MapAsyncBenchmark.mapAsyncUnordered               1     true  thrpt    5    585846,883 ±  41347,242  ops/s
[info] MapAsyncBenchmark.mapAsyncUnordered1              1    false  thrpt    5  19139110,394 ±  37640,603  ops/s
[info] MapAsyncBenchmark.mapAsyncUnordered1              1     true  thrpt    5    561255,268 ±  98371,076  ops/s
```

Here's the changes I did to easily test the impls against each other:
[bench.diff.txt](https://github.com/akka/akka/files/9567671/bench.diff.txt)