@He-Pin 2022-09-15
![image](https://user-images.githubusercontent.com/501740/190321842-2d9ba3db-1288-4a2d-8684-446215532027.png)

with power plugged and performance mode