@johanandren 2022-09-15
Allright, strange that I couldn't repeat on arm64, I'll give it a run on a Linux Intel machine as well and see if I can repeat there.