@johanandren 2022-09-15
On a i5/ubuntu I can see an improvement as well, (but not as large a gap as for you):

```
[info] Benchmark                                (spawn)   Mode  Cnt         Score        Error  Units
[info] MapAsyncOneBenchmark.mapAsyncUnordered     false  thrpt   20  19796075.599 ±  30653.944  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered      true  thrpt   20    591048.785 ±   1394.243  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1    false  thrpt   20  22572466.943 ± 160190.642  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1     true  thrpt   20    576528.865 ±   1866.207  ops/s
```

Would be good to profile to get an understanding if it in fact is the few integer ops are indeed the cause (or if there is something completely unexpected that shouldn't be happening in the original impl)