@MarchLiu 2022-09-15
RMBP MacBook Pro (15-inch, 2017) Intel Core i7 x4
```
[info] 229296.890 ops/s
[info] Result "akka.stream.MapAsyncOneBenchmark.mapAsyncUnordered1":
[info]   265067.736 ±(99.9%) 29827.374 ops/s [Average]
[info]   (min, avg, max) = (218417.932, 265067.736, 350012.593), stdev = 34349.248
[info]   CI (99.9%): [235240.363, 294895.110] (assumes normal distribution)
[info] # Run complete. Total time: 00:27:14
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                (spawn)   Mode  Cnt         Score         Error  Units
[info] MapAsyncOneBenchmark.mapAsyncUnordered     false  thrpt   20   5322055.668 ± 1157358.214  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered      true  thrpt   20    253088.287 ±   24018.281  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1    false  thrpt   20  15192007.522 ±   50890.876  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1     true  thrpt   20    265067.736 ±   29827.374  ops/s
[success] Total time: 1850 s (30:50), completed 2022年9月15日 下午11:10:56
```