@MarchLiu 2022-09-15
Ubuntu Linux  Intel(R) Xeon(R) Bronze 3204 CPU @ 1.90GHz x6
```
[info] 123476.716 ops/s
[info] Result "akka.stream.MapAsyncOneBenchmark.mapAsyncUnordered1":
[info]   121379.404 ±(99.9%) 1471.038 ops/s [Average]
[info]   (min, avg, max) = (117772.629, 121379.404, 123476.716), stdev = 1694.050
[info]   CI (99.9%): [119908.366, 122850.442] (assumes normal distribution)
[info] # Run complete. Total time: 00:27:35
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                (spawn)   Mode  Cnt        Score        Error  Units
[info] MapAsyncOneBenchmark.mapAsyncUnordered     false  thrpt   20  3413311.547 ± 185446.572  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered      true  thrpt   20   112378.554 ±   2745.039  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1    false  thrpt   20  4059178.970 ±  49648.912  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1     true  thrpt   20   121379.404 ±   1471.038  ops/s
[success] Total time: 1785 s (29:45), completed Sep 15, 2022, 11:23:43 PM
```