@gcnyin 2022-09-16
macbook m1 pro, openjdk 1.8.0_342

```
[info] Benchmark                                (spawn)   Mode  Cnt         Score         Error  Units
[info] MapAsyncOneBenchmark.mapAsyncUnordered     false  thrpt   20  18453380.142 ± 1085170.728  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered      true  thrpt   20    514386.083 ±   34420.446  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1    false  thrpt   20  19405036.958 ±  724018.675  ops/s
[info] MapAsyncOneBenchmark.mapAsyncUnordered1     true  thrpt   20    601906.826 ±   43473.245  ops/s
```