@patriknw 2022-09-16
I see a speedup for the spawn false case of 20%.
For spawn true it's a small improvement, but not much.

@johanandren do you think this changes the "detached" semantics for already completed futures compared to existing implementation? Would that be ok or could it introduce any surprises?