@johanandren 2022-09-16
I don't see that it is more detached than the existing MapAsyncUnordered stage. Both pull eagerly after push rathe rthan wait for pull (and still not sure what explains the performance improvement, I'd expect it to be closer to the same looking at the paths that should be exercised by parallelism 1 - no going through the buffer there)